remove odd classes refactoring inventory search service
