+++ VirtoCommerce.InventoryModule.Data/packages.config
@@ -9,7 +9,7 @@
   <package id="Owin" version="1.0" targetFramework="net461" />
   <package id="StackExchange.Redis" version="1.2.6" targetFramework="net461" />
   <package id="valueinjecter" version="2.3.3" targetFramework="net451" />
-  <package id="VirtoCommerce.Domain" version="2.25.0" targetFramework="net461" />
-  <package id="VirtoCommerce.Platform.Core" version="2.13.26" targetFramework="net461" />
+  <package id="VirtoCommerce.Domain" version="2.25.30" targetFramework="net461" />
+  <package id="VirtoCommerce.Platform.Core" version="2.13.42" targetFramework="net461" />
   <package id="VirtoCommerce.Platform.Data" version="2.13.26" targetFramework="net461" />
 </packages>
