+++ VirtoCommerce.InventoryModule.Test/packages.config
@@ -8,8 +8,8 @@
   <package id="Newtonsoft.Json" version="11.0.2" targetFramework="net461" />
   <package id="System.Threading.Tasks.Extensions" version="4.3.0" targetFramework="net461" />
   <package id="System.ValueTuple" version="4.4.0" targetFramework="net461" />
-  <package id="VirtoCommerce.Domain" version="2.25.0" targetFramework="net461" />
-  <package id="VirtoCommerce.Platform.Core" version="2.13.26" targetFramework="net461" />
+  <package id="VirtoCommerce.Domain" version="2.25.30" targetFramework="net461" />
+  <package id="VirtoCommerce.Platform.Core" version="2.13.42" targetFramework="net461" />
   <package id="xunit" version="2.3.1" targetFramework="net461" />
   <package id="xunit.abstractions" version="2.0.1" targetFramework="net461" />
   <package id="xunit.analyzers" version="0.7.0" targetFramework="net461" />
