+++ VirtoCommerce.InventoryModule.Web/packages.config
@@ -12,9 +12,9 @@
   <package id="StackExchange.Redis" version="1.2.6" targetFramework="net461" />
   <package id="Unity" version="4.0.1" targetFramework="net451" />
   <package id="valueinjecter" version="2.3.3" targetFramework="net451" />
-  <package id="VirtoCommerce.Domain" version="2.25.0" targetFramework="net461" />
+  <package id="VirtoCommerce.Domain" version="2.25.30" targetFramework="net461" />
   <package id="VirtoCommerce.Module" version="0.17.0" targetFramework="net461" developmentDependency="true" />
-  <package id="VirtoCommerce.Platform.Core" version="2.13.26" targetFramework="net461" />
+  <package id="VirtoCommerce.Platform.Core" version="2.13.42" targetFramework="net461" />
   <package id="VirtoCommerce.Platform.Core.Web" version="2.13.26" targetFramework="net461" />
   <package id="VirtoCommerce.Platform.Data" version="2.13.26" targetFramework="net461" />
 </packages>
